feat(ecs): add Timer class and change elapsed time type from 'float' to 'uint64_t'

+++ GameEngine/Systems/ASystem.hpp
@@ -44,7 +44,7 @@ class ASystem
 
         virtual ~ASystem() = default;
 
-        virtual void update(RegistryManager &manager, std::shared_ptr<AGraphical> &graphical, float elapsedMs) = 0;
+        virtual void update(RegistryManager &manager, std::shared_ptr<AGraphical> &graphical, uint64_t elapsedMs) = 0;
 };
 
 #endif /* !ASYSTEM_HPP_ */
